% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{addKnot.gaussian}
\alias{addKnot.gaussian}
\title{Add a new knot and use a Metropolis-Hastings step
to determine if we keep the changes to the model}
\usage{
addKnot.gaussian(model.options, knots.previous, outcomes, times.dropout,
  times.observation, covariates, X.previous, Theta.previous, Z, alpha,
  betaCovariates, sigma.error)
}
\description{
Add a new knot and use a Metropolis-Hastings step
to determine if we keep the changes to the model
}
\examples{

}

