% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{bayes.splines.fit}
\alias{bayes.splines.fit}
\title{bayes.splines.fit}
\usage{
bayes.splines.fit(model.options, dist, groups, covariates.var, iterations)
}
\arguments{
\item{model.options}{the original model options}

\item{dist}{the distribution of the outcome ("gaussian" or "binary")}

\item{groups}{the list of groups}

\item{covariates.var}{list of covariate column names}

\item{iterations}{the model run iterations after removing burn-in iterations and thinning}
}
\description{
Model fit for a Bayesian spline model run
}
