% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{bayes.splines.iteration}
\alias{bayes.splines.iteration}
\title{bayes.splines.iteration}
\usage{
bayes.splines.iteration(knots = NULL, Theta = NULL,
  betas.covariates = NULL, sigma.residual = 1, sigma.randomIntercept = 1,
  sigma.randomSlope = 1, sigma.randomInterceptSlope = 0, sigma.error = 1,
  sigma.error.shape = 0.001, sigma.error.rate = 0.001)
}
\arguments{
\item{knots}{list of knot positions by group}

\item{Theta}{list of spline coefficients by group}

\item{betas.covariates}{regression coefficients related to covariates}

\item{sigma.residual}{covariance of the spline coefficients}

\item{sigma.randomIntercept}{variance of the random intercepts}

\item{sigma.randomSlope}{variance of the random slopes}

\item{sigma.randomInterceptSlope}{covariance of the random intercepts and slopes}

\item{sigma.error}{for Gaussian outcomes, the residual error}

\item{sigma.error.shape}{for Gaussian outcomes, the shape hyperparamter of the inverse gamma
distribution of the residual error}

\item{sigma.error.rate}{for Gaussian outcomes, the rate hyperparamter of the inverse gamma
distribution of the residual error}

\item{shape.tau}{}
}
