% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{bayes.splines.model.options}
\alias{bayes.splines.model.options}
\title{Simulation and model options for the natural b-spline model}
\usage{
bayes.splines.model.options(iterations = 10000, burnin = 500, thin = NA,
  knots.prob.birth = 0.5, knots.min = 1, knots.max = NA,
  knots.positions.start = NULL, knots.positions.candidate = NULL,
  prob.min = 1e-05, prob.max = 0.99999, dropout.estimationTimes = NULL,
  shape.tau = NULL, rate.tau = NULL, sigma.beta = NULL,
  lambda.numKnots = NULL, sigma.error = 0, sigma.error.df = NULL,
  sigma.error.scale = NULL, eta.null = NULL)
}
\arguments{
\item{iterations}{number of iterations for the MCMC simulation}

\item{burnin}{burn in period for the simulation, i.e. the number of 
iterations to throw away at the beginning of the simulation}

\item{thin}{thinning interval, i.e. if thin=n, only keep every nth iteration}

\item{knots.prob.birth}{probability of adding a new knot to the model on a given iteration}

\item{knots.min}{minimum number of knots in the model. Must be greater than or equal to 1.}

\item{knots.max}{maximum number of knots in the model.}
}
\description{
Simulation and model options for the natural b-spline model
}

