% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{calculateDropoutTimeSpecificSlope}
\alias{calculateDropoutTimeSpecificSlope}
\title{Calculate the estimated slope at the given dropout times}
\usage{
calculateDropoutTimeSpecificSlope(dropoutEstimationTimes, knotsByGroup,
  ThetaByGroup, subjectsPerGroup, times.observation, times.dropout)
}
\arguments{
\item{dropoutEstimationTimes}{list of times at which to estimate the slope}

\item{knotsByGroup}{list of knots by group}

\item{ThetaByGroup}{list of spline coefficients by group}

\item{subjectsPerGroup}{list of number of subjects by group}

\item{times.observation}{observation times}

\item{times.dropout}{dropout times}
}
\description{
Calculate the estimated slope at the given dropout times
}

