% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{calculateMarginalSlope}
\alias{calculateMarginalSlope}
\title{Calculate the marginal slope at the given iteration}
\usage{
calculateMarginalSlope(knotsByGroup, ThetaByGroup, subjectsPerGroup,
  times.dropout)
}
\arguments{
\item{knotsByGroup}{list of knots by group}

\item{ThetaByGroup}{list of spline coefficients by group}

\item{subjectsPerGroup}{list of number of subjects by group}

\item{times.dropout}{dropout times}
}
\description{
Calculate the marginal slope at the given iteration
}
