% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichletProcessFit.R
\name{dirichlet.iteration}
\alias{dirichlet.iteration}
\title{dirichlet.iteration}
\usage{
dirichlet.iteration(weights.mixing = NULL, weights.conditional = NULL,
  cluster.assignments = NULL, betas = NULL, betas.deviations = NULL,
  betas.covariates = NULL, betas.covariates.mu = NULL,
  betas.covariates.sigma = NULL, dp.dist.mu0 = NULL,
  dp.dist.sigma0 = NULL, dp.cluster.sigma = NULL,
  dp.concentration = NULL, cluster.N = NULL, cluster.mu = NULL,
  sigma.error = NULL, expected.intercept = NULL,
  expected.slope = NULL, slope.dropoutTimes = NULL,
  density.intercept = NULL, density.slope = NULL)
}
\arguments{
\item{weights.mixing}{vector containing the probability of belonging to cluster k, 
for k = 1 to the number of clusters}

\item{weights.conditional}{vector containing the probability of belonging to cluster k, given
that the subject was not in clusters 1 to k - 1}

\item{cluster.assignments}{current cluster assignments}

\item{betas}{A (k x 3) matrix of regression coefficients for the random intercept, slope,
and log of the dropout time for each cluster, with k = number of clusters}

\item{betas.deviations}{An (N x k) matrix of subject specific deviations from the cluster means}

\item{betas.covariates}{A (c x 1) vector of regression coefficients for covariates, 
with c = number of covariates}

\item{betas.covariates.mu}{A (c x 1) vector representing the mean of the distribution of 
regression coefficients related to covariates}

\item{betas.covariates.sigma}{A (c x c) vector representing the covariance of the distribution of 
regression coefficients related to covariates}

\item{dp.dist.mu0}{A (3 x 1) vector of means for the baseline distribution of the Dirichlet process}

\item{dp.dist.sigma0}{A (3 x 3) matrix representing the covariance of the baseline 
distribution of the Dirichlet process}

\item{dp.cluster.sigma}{A (3 x 3) matrix representing the covariance of the random intercept, slope,
and log dropout time for each cluster.  This covariance is the same for each cluster}

\item{dp.concentration}{the concentration parameter (alpha) for the Dirichlet process}

\item{cluster.N}{A (k x 1) vector indicating the number of subjects in cluster k, for k = 1 to the
number of clusters}

\item{cluster.mu}{A (k x 3) matrix of means for the random intercept, slope, and log dropout time
in cluster k, for k = 1 to the number of clusters}

\item{sigma.error}{For Gaussian outcomes, the residual error variance}

\item{expected.intercept}{the expected value of the random intercept}

\item{expected.slope}{the expected value of the random slope}

\item{slope.dropoutTimes}{estimated slopes by dropout times}

\item{density.intercept}{estimated density of the random intercepts}

\item{density.slope}{estimated density of the random slopes}

\item{expected.intercept}{expected value of the random intercepts}

\item{expected.slope}{expected value of the random slopes}
}
\description{
dirichlet.iteration
}
