% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichletProcessFit.R
\name{dirichlet.model.options}
\alias{dirichlet.model.options}
\title{dirichlet.model.options}
\usage{
dirichlet.model.options(iterations = 10000, burnin = 500, thin = 1,
  print = 1, start.weights.mixing = NULL, n.clusters = 15,
  dropout.estimationTimes = NULL, dropout.offset = 0,
  dp.concentration = NULL, dp.concentration.alpha = NULL,
  dp.concentration.beta = NULL, dp.cluster.sigma = NULL,
  dp.cluster.sigma.nu0 = NULL, dp.cluster.sigma.T0 = NULL,
  dp.dist.mu0 = NULL, dp.dist.mu0.mb = NULL, dp.dist.mu0.Sb = NULL,
  dp.dist.sigma0 = NULL, dp.dist.sigma0.nub = NULL,
  dp.dist.sigma0.Tb = NULL, betas.covariates = NULL,
  betas.covariates.mu = NULL, betas.covariates.sigma = NULL,
  sigma.error = NULL, sigma.error.tau = NULL,
  sigma.error.perGroup = FALSE)
}
\arguments{
\item{iterations}{number of iterations for the MCMC simulation}

\item{burnin}{burn in period for the simulation, i.e. the number of 
iterations to throw away at the beginning of the simulation}

\item{thin}{thinning interval, i.e. if thin=n, only keep every nth iteration}

\item{print}{printing interval, i.e. if print=n, print a counter on every nth iteration}

\item{dp.concentration}{prior value for the concentration parameter of the Dirichlet process}

\item{dp.concentration.alpha}{Shape parameter for the hyperprior Gamma distribution of the
concentration parameter of the Dirichlet process}

\item{dp.concentration.beta}{Rate parameter for the hyperprior Gamma distribution of the
concentration parameter of the Dirichlet process}

\item{dp.cluster.sigma}{prior for the common cluster covariance}

\item{dp.cluster.sigma.nu0}{Degrees of freedom for the hyperprior inverse Wishart distribution of the
common cluster covariance}

\item{dp.cluster.sigma.T0}{Scale matrix for the hyperprior inverse Wishart distribution of the
common cluster covariance}

\item{dp.dist.mu0}{prior mean for the baseline distribution of the Dirichlet process}

\item{dp.dist.mu0.mb}{Mean for the hyperprior Gaussian distribution of the mean of
the baseline distribution of the Dirichlet process}

\item{dp.dist.mu0.Sb}{Covariance for the hyperprior Gaussian distribution of the mean of
the baseline distribution of the Dirichlet process}

\item{dp.dist.sigma0}{prior covariance for the baseline distribution of the Dirichlet process}

\item{dp.dist.sigma0.nub}{Degrees of freedom for the hyperprior inverse Wishart distribution of 
the covariance of the baseline distribution of the Dirichlet process}

\item{dp.dist.sigma0.Tb}{Scale matrix for the hyperprior inverse Wishart distribution of 
the covariance of the baseline distribution of the Dirichlet process}

\item{betas.covariates}{prior value for covariate regression coefficients}

\item{betas.covariates.mu}{Prior mean for the covariate regression coefficients}

\item{betas.covariates.sigma}{Prior covariance for the covariate regression coefficients}

\item{sigma.error}{Prior for the residual error (Gaussian outcomes only)}

\item{sigma.error.tau}{Hyperprior for the residual error (Gaussian outcomes only)}

\item{sigma.error.perGroup}{If TRUE, separate sigma error parameters will be fit per group}

\item{numClusters}{number of clusters for the Dirichlet Process stick breaking model}
}
\description{
Model options for the Dirichlet Process model.  Includes starting values, priors and
simulation/MCMC parameters.
}

