% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informativeDropout.R
\name{getInitialEstimatesCovariates}
\alias{getInitialEstimatesCovariates}
\title{Fit a simple linear model to get initial values for regression
coefficients associated with covariates}
\usage{
getInitialEstimatesCovariates(dist, covariates, outcomes)
}
\arguments{
\item{dist}{the distribution of the outcome ("gaussian" or "binary")}

\item{covariates}{date frame containing covariate values}

\item{outcomes}{vector of outcomes}
}
\description{
Fit a simple linear model to get initial values for regression
coefficients associated with covariates
}

