% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichletProcessFit.R
\name{informativeDropout.bayes.dirichlet}
\alias{informativeDropout.bayes.dirichlet}
\title{infortmativeDropout.bayes.dirichlet}
\usage{
informativeDropout.bayes.dirichlet(data, ids.var, outcomes.var, groups.var,
  covariates.var, times.dropout.var, times.observation.var, censoring.var,
  dist, model.options)
}
\arguments{
\item{data}{the data set}

\item{ids.var}{column of the data set containing participant identifiers}

\item{outcomes.var}{column of the data set containing the outcome variable}

\item{groups.var}{column of the data set indicating the treatment groups}

\item{covariates.var}{list of columns in the data set containing covariates}

\item{times.dropout.var}{column of the data set containing dropout times}

\item{times.observation.var}{column of the data set containing}

\item{dist}{the distribution of the outcome, valid values are "gaussian" or "binary"}

\item{model.options}{model options (see dirichlet.model.options for details)}
}
\description{
Fit a bayesian model which accounts for dropout by modeling the 
releationship between dropout time and slope with natural cubic B-splines
}
