% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedModelFit.R
\name{mixed.model.options}
\alias{mixed.model.options}
\title{Model options for the mixed model fit.  A convenience wrapper
around lme}
\usage{
mixed.model.options(knots.positions = NULL)
}
\arguments{
\item{knots.positions}{knot positions}
}
\description{
Model options for the mixed model fit.  A convenience wrapper
around lme
}
