% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichletProcessFit.R
\name{plot.density.dirichlet.fit}
\alias{plot.density.dirichlet.fit}
\title{plot.density.dirichlet.fit}
\usage{
\method{plot}{density.dirichlet.fit}(fit, name = "slope")
}
\description{
density plot for a parameter
}

