% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{plot.slopeByDropout.bayes.splines.fit}
\alias{plot.slopeByDropout.bayes.splines.fit}
\title{plot.slopeByDropout.bayes.splines.fit}
\usage{
\method{plot}{slopeByDropout.bayes.splines.fit}(fit, groups = NULL,
  xlim = NULL, ylim = NULL, show_ci = TRUE, overlay = TRUE,
  lower_tail = 0.025, upper_tail = 0.975, colors = NULL)
}
\arguments{
\item{fit}{the Dirichlet model fit}

\item{groups}{list of groups to include in the plot(s)}

\item{xlim}{the limits of the X-axis}

\item{ylim}{the limits of the Y-axis}

\item{show_ci}{if TRUE, displays the upper and lower credible interval}

\item{overlay}{if TRUE, overlays all groups on the same plot}

\item{lower_tail}{the lower tail probability for the credible interval}

\item{upper_tail}{the upper tail probability for the credible interval}

\item{colors}{vector of colors for each group}
}
\description{
Plot the slope by dropout time for a Bayesian spline fit
}

