% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.slopeByDropout}
\alias{plot.slopeByDropout}
\title{plot the slope by dropout time}
\usage{
plot.slopeByDropout(fit, ...)
}
\arguments{
\item{fit}{the model fit object}
}
\description{
plot the slope by dropout time
}

