% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{plot.trace.bayes.splines.fit}
\alias{plot.trace.bayes.splines.fit}
\title{Create a trace plot of the specified parameter}
\usage{
\method{plot}{trace.bayes.splines.fit}(fit, type = "marginal_slope",
  groups = NULL, params = NULL)
}
\arguments{
\item{fit}{bayes.splines.fit object from an informativeDropout run.}
}
\description{
Create a trace plot of the specified parameter
}

