% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{prob.acceptance}
\alias{prob.acceptance}
\title{calculate the acceptance probability for a given update step in the Bayesian
spline model run.}
\usage{
prob.acceptance(fit, action)
}
\arguments{
\item{fit}{the bayes.splines.fit object}

\item{action}{the update action.  Valid values are "knot.add", "knot.remove",
"knot.move", "fixedEffects", "fixedEffectsCovariates"}
}
\description{
calculate the acceptance probability for a given update step in the Bayesian
spline model run.
}
