% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{sensitivity.bayes.splines.fit}
\alias{sensitivity.bayes.splines.fit}
\title{sensitivity.bayes.splines.fit}
\usage{
\method{sensitivity}{bayes.splines.fit}(fit, times.estimation, deltas,
  data.onePerSubject, times.dropout.var, group.var = NULL,
  covariates.time.var = NULL, covariates.nontime.var = NULL)
}
\arguments{
\item{fit}{the fit object from a call to informativeDropout, using the 'bayes.splines' model}

\item{times.estimation}{vector of times at which to estimate the expected value of the outcome}

\item{deltas}{vector of multipliers to increase/decrease slope after dropout}

\item{data.onePerSubject}{data frame containing one row per participant. The data frame
must include dropout time and grouping variable (if multi-group design). Both time- and non-time-interacted
covariates should be specified if the user wishes to include them in the sensitivity analysis.}

\item{times.dropout.var}{column name containing drop out times}

\item{group.var}{column name containing group values}

\item{covariates.time.var}{vector of columns containing time-interacted covariates}

\item{covariates.nontime.var}{vector of columns containing non-time-interacted covariates}
}
\description{
Sensitivity analysis for models using the 'bayes.splines' model  Allows users to
determine how sensitive the results are if the true slope after dropout changes.
}
