% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_1}
\alias{sim_1}
\title{A simulated data set based on the Multicenter Aids study
Longitudinal data of HIV in men who are both users and non-users of IV drugs}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\usage{
sim_1
}
\description{
A simulated data set based on the Multicenter Aids study
Longitudinal data of HIV in men who are both users and non-users of IV drugs
}
\keyword{datasets}

