% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{summary.bayes.splines.fit}
\alias{summary.bayes.splines.fit}
\title{Summarize a Bayesian spline model run}
\usage{
\method{summary}{bayes.splines.fit}(fit)
}
\arguments{
\item{fit}{bayes.splines.fit object from an informativeDropout run.}
}
\description{
Summarize a Bayesian spline model run
}
