% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichletProcessFit.R
\name{summary.dirichlet.fit}
\alias{summary.dirichlet.fit}
\title{summary.dirichlet.fit}
\usage{
\method{summary}{dirichlet.fit}(fit)
}
\arguments{
\item{fit}{dirichlet.fit object from an informativeDropout run.}
}
\description{
Summarize a Dirichlet model run
}

