% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{summary.sensitivity.bayes.splines.fit}
\alias{summary.sensitivity.bayes.splines.fit}
\title{summary.sensitivity.bayes.splines.fit}
\usage{
\method{summary}{sensitivity.bayes.splines.fit}(sensitivity.fit,
  tail.lower = 0.025, tail.upper = 0.975)
}
\arguments{
\item{sensitivity.fit}{the fit object from a call to sensitivity for a 'bayes.splines' model}

\item{tail.lower}{lower tail for confidence bounds}

\item{tail.upper}{upper tail for confidence bounds}
}
\description{
Summarize sensitivity analysis results for a Bayesian spline model
}
