% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{untreated_hiv}
\alias{untreated_hiv}
\title{Longitudinal data of HIV in men who are both users and non-users of IV drugs}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\usage{
untreated_hiv
}
\description{
Longitudinal data of HIV in men who are both users and non-users of IV drugs
}
\keyword{datasets}
