% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{updateCovarianceParameters.gaussian}
\alias{updateCovarianceParameters.gaussian}
\title{Add a new knot and use a Metropolis-Hastings step
to determine if we keep the changes to the model}
\usage{
updateCovarianceParameters.gaussian(model.options, totalObservations,
  numSubjects, firstObsPerSubject, outcomes, covariates, X, Theta, Z,
  alpha, betaCovariates, sigma.error)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add a new knot and use a Metropolis-Hastings step
to determine if we keep the changes to the model
}
\examples{
add(1, 1)
add(10, 1)

}
