% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{updateFixedEffectsCovariates.gaussian}
\alias{updateFixedEffectsCovariates.gaussian}
\title{Update the regression coefficients related to common
covariates and group effects}
\usage{
updateFixedEffectsCovariates.gaussian(model.options, outcomes, covariates, X,
  Theta, Z, alpha, betaCovariates.previous, sigma.error)
}
\description{
Update the regression coefficients related to common
covariates and group effects
}
