% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{updateRandomEffects.gaussian}
\alias{updateRandomEffects.gaussian}
\title{Update the random effects}
\usage{
updateRandomEffects.gaussian(numSubjects, numObservations, firstObsPerSubject,
  subjectsPerGroup, ids, outcomes, times.observation, covariates, X, Theta, Z,
  alpha, betaCovariates, sigma.randomIntercept, sigma.randomSlope,
  sigma.randomInterceptSlope, sigma.error)
}
\description{
Update the random effects
}
\examples{

}
