% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{addKnot.binary}
\alias{addKnot.binary}
\title{For binary outcomes, add a new knot and use a Metropolis-Hastings step
to determine if we keep the changes to the model.}
\usage{
addKnot.binary(model.options, knots.previous, outcomes, times.dropout,
  times.observation, covariates, X.previous, Theta.previous, Z, alpha,
  betaCovariates)
}
\description{
For binary outcomes, add a new knot and use a Metropolis-Hastings step
to determine if we keep the changes to the model.
}

