% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{addKnot.gaussian}
\alias{addKnot.gaussian}
\title{addKnot.gaussian}
\usage{
addKnot.gaussian(model.options, knots.previous, knots.positions.candidate,
  outcomes, times.dropout, times.observation, covariates, X.previous,
  Theta.previous, Z, alpha, betaCovariates, sigma.error)
}
\arguments{
\item{model.options}{model options}

\item{knots.previous}{previous set of knots}

\item{outcomes}{vector of outcomes}

\item{times.dropout}{vector of dropout times}

\item{times.observation}{vector of observation times}

\item{covariates}{data frame of covariates}

\item{X.previous}{previous X matrix}

\item{Theta.previous}{previous Theta values}

\item{Z}{random effects design matrix}

\item{alpha}{random effects}

\item{betaCovariates}{regression coefficients for covariates}

\item{sigma.error}{residual variance}
}
\value{
list containing updated X, knots, Theta, and boolean indicating if change accepted
}
\description{
Add a new knot and use a Metropolis-Hastings step
to determine if we keep the changes to the model
}

