% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{calculateMarginalSlope}
\alias{calculateMarginalSlope}
\title{Calculate the marginal slope at the given iteration}
\usage{
calculateMarginalSlope(knotsByGroup, ThetaByGroup, subjectsPerGroup,
  times.dropout)
}
\description{
Calculate the marginal slope at the given iteration
}

