% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{getInitialEstimatesTheta}
\alias{getInitialEstimatesTheta}
\title{Fit a simple linear model to get initial values for regression
coefficients associated with splines}
\usage{
getInitialEstimatesTheta(dist, groupList, X, outcomes)
}
\arguments{
\item{dist}{the distribution of the outcome ("gaussian" or "binary")}

\item{outcomes}{vector of outcomes}

\item{covariates}{date frame containing covariate values}
}
\description{
Fit a simple linear model to get initial values for regression
coefficients associated with splines
}

