% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getInitialEstimatesTheta}
\alias{getInitialEstimatesTheta}
\title{Fit a simple linear model to get initial values for regression
coefficients associated with splines}
\usage{
getInitialEstimatesTheta(dist, groupList, X, outcomes)
}
\arguments{
\item{dist}{the distribution of the outcome ("gaussian" or "binary")}

\item{groupList}{list of groups}

\item{X}{data frame containing predictor variables (assumes a column of ones for an intercept)}

\item{outcomes}{vector of outcomes}
}
\description{
Fit a simple linear model to get initial values for regression
coefficients associated with splines
}
