% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{informativeDropout.bayes.splines}
\alias{informativeDropout.bayes.splines}
\title{Fit a varying coefficient model for longitudinal studies with
informative dropout.  Uses a Bayesian approach with a spline fit
to model the relationship between dropout times and slope}
\usage{
informativeDropout.bayes.splines(data, ids.var, outcomes.var, groups.var,
  covariates.var, times.dropout.var, times.observation.var, dist, model.options)
}
\description{
Fit a varying coefficient model for longitudinal studies with
informative dropout.  Uses a Bayesian approach with a spline fit
to model the relationship between dropout times and slope
}

