% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informativeDropout.R
\name{informativeDropout}
\alias{informativeDropout}
\title{Fit a varying coefficient model for longitudinal studies with
informative dropout.}
\usage{
informativeDropout(data, ids.var, outcomes.var, groups.var, covariates.var,
  times.dropout.var, times.observation.var, method = "bayes.splines",
  dist = "normal", model.options)
}
\description{
Fit a varying coefficient model for longitudinal studies with
informative dropout.
}
\examples{

}

