% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichletProcessFit.R
\name{plot.density.dirichlet.fit}
\alias{plot.density.dirichlet.fit}
\title{plot.density.dirichlet.fit}
\usage{
\method{plot}{density.dirichlet.fit}(fit, params = NULL, groups = NULL,
  overlay = TRUE, xlim = NULL, ylim = NULL, colors = NULL, lty = 1)
}
\arguments{
\item{fit}{the Dirichlet model fit}

\item{params}{the vector indicating the values to plot (either "density.intercept" or
"density.slope")}

\item{groups}{list of groups to include in the plot(s)}

\item{xlim}{the limits of the X-axis}

\item{ylim}{the limits of the Y-axis}

\item{colors}{vector of colors for each group}
}
\description{
Plot the densities of the random effects for a Dirichlet Process model
}

