% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichletProcessFit.R
\name{plot.slopeByDropout.dirichlet.fit}
\alias{plot.slopeByDropout.dirichlet.fit}
\title{plot.slopeByDropout.dirichlet.fit}
\usage{
\method{plot}{slopeByDropout.dirichlet.fit}(fit, group = 1, xlim = NULL,
  ylim = NULL, lower_tail = 0.025, upper_tail = 0.975)
}
\description{
plot the slope by dropout time
}

