% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{removeKnot.gaussian}
\alias{removeKnot.gaussian}
\title{Remove a knot and use a Metropolis-Hastings step
to determine if we keep the changes to the model}
\usage{
removeKnot.gaussian(model.options, knots.previous, outcomes, times.dropout,
  times.observation, covariates, X.previous, Theta.previous, Z, alpha,
  betaCovariates, sigma.error)
}
\description{
Remove a knot and use a Metropolis-Hastings step
to determine if we keep the changes to the model
}
\examples{

}

