% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sensitivity}
\alias{sensitivity}
\title{perform sensitivity analysis on the slope results}
\usage{
sensitivity(fit, times.estimation = NULL, deltas = NULL)
}
\arguments{
\item{fit}{the model fit object}
}
\description{
perform sensitivity analysis on the slope results
}

