% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{updateFixedEffectsCovariates.binary}
\alias{updateFixedEffectsCovariates.binary}
\title{Update the regression coefficients related to common
covariates and group effects}
\usage{
updateFixedEffectsCovariates.binary(model.options, outcomes, covariates, X,
  Theta, Z, alpha, betaCovariates.previous)
}
\description{
Update the regression coefficients related to common
covariates and group effects
}
