% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianSplineFit.R
\name{updateFixedEffectsCovariates.gaussian}
\alias{updateFixedEffectsCovariates.gaussian}
\title{Update the regression coefficients related to common
covariates and group effects}
\usage{
updateFixedEffectsCovariates.gaussian(outcomes, covariates, X, Theta, Z, alpha,
  betaCovariates.previous, sigma.error, sigma.beta)
}
\description{
Update the regression coefficients related to common
covariates and group effects
}

